refactor(random): Refactored functions for crushing seeding sequence to a single value to a more uniform API.
fix(random): Fixed repeated values in the initial state of some PRNGs when initialized with a seeding sequence. This may lead to a change in generated sequences for some existing code, however sequence stability is only guaranteed across different environments, not different library versions, so this is not considered a breaking change for the purposes of semver.